+++ Back_End/src/main/java/com/ssafy/youniverse/dto/req/ActorReqDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ActorReqDto {
-    private Integer actorId;
+    private Long actorId;
     private String actorName;
     private String actorImage;
 }
