+++ Back_End/src/main/java/com/ssafy/youniverse/dto/req/MemberRecommendMovieReqDto.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Setter
 public class MemberRecommendMovieReqDto {
     private String email;
-    private List<Integer> movieIdList;
+    private List<Long> movieIdList;
 }
