+++ Back_End/src/main/java/com/ssafy/youniverse/dto/req/MemberReqDto.java
@@ -8,12 +8,12 @@ import java.util.List;
 @Getter
 @Setter
 public class MemberReqDto {
-    private Integer memberId;
+    private Long memberId;
     private String nickname;
     private String email;
     private String gender;
     private Byte age;
     private String introduce;
-    private List<Integer> ottList; //회원과 관련된 Ott 식별자
-    private List<Integer> keywordList; //회원과 관련된 Keyword 식별자
+    private List<Long> ottList; //회원과 관련된 Ott 식별자
+    private List<Long> keywordList; //회원과 관련된 Keyword 식별자
 }
