+++ Back_End/src/main/java/com/ssafy/youniverse/dto/res/YoutubeKeywordResDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class YoutubeKeywordResDto {
-    private Integer youtubeKeywordId;
+    private Long youtubeKeywordId;
     private String youtubeKeywordName;
     private Byte movieRank;
 }
