+++ Back_End/src/main/java/com/ssafy/youniverse/entity/Follow.java
@@ -15,7 +15,7 @@ public class Follow extends Auditable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer followId;
+    private Long followId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "follower_id")
