+++ Back_End/src/main/java/com/ssafy/youniverse/mapper/RecommendMovieMapper.java
@@ -11,8 +11,8 @@ import java.util.List;
 @Mapper(componentModel = "spring")
 public interface RecommendMovieMapper extends CustomMapper {
     @Mapping(source = "movieId", target = "movie.movieId")
-    RecommendMovie movieIdToRecommendMovie(Integer movieId);
-    List<RecommendMovie> movieIdListToRecommendMovies(List<Integer> movieIdList);
+    RecommendMovie movieIdToRecommendMovie(Long movieId);
+    List<RecommendMovie> movieIdListToRecommendMovies(List<Long> movieIdList);
 
     @Mapping(source = "recommendMovie.member", target = "memberSimpleResDto")
     @Mapping(source = "recommendMovie.movie", target = "movieSimpleResDto")
