+++ Back_End/src/main/java/com/ssafy/youniverse/repository/ActorRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.Actor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActorRepository extends JpaRepository<Actor, Integer> {
+public interface ActorRepository extends JpaRepository<Actor, Long> {
 }
