+++ Back_End/src/main/java/com/ssafy/youniverse/repository/DirectorRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.Director;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DirectorRepository extends JpaRepository<Director, Integer> {
+public interface DirectorRepository extends JpaRepository<Director, Long> {
 }
