+++ Back_End/src/main/java/com/ssafy/youniverse/repository/KeywordMovieRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.KeywordMovie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KeywordMovieRepository extends JpaRepository<KeywordMovie, Integer> {
+public interface KeywordMovieRepository extends JpaRepository<KeywordMovie, Long> {
 }
