+++ Back_End/src/main/java/com/ssafy/youniverse/repository/KeywordRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface KeywordRepository extends JpaRepository<Keyword, Integer>, KeywordRepositoryCustom {
+public interface KeywordRepository extends JpaRepository<Keyword, Long>, KeywordRepositoryCustom {
     Optional<Keyword> findByKeywordName(String keywordName);
 }
