+++ Back_End/src/main/java/com/ssafy/youniverse/repository/MovieRepositoryCustom.java
@@ -5,5 +5,5 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 public interface MovieRepositoryCustom {
-    Page<Movie> findAllQueryDsl(Pageable pageable, Integer memberId, String director, String actor, String title, Integer type, Integer ottId);
+    Page<Movie> findAllQueryDsl(Pageable pageable, Long memberId, String director, String actor, String title, Integer type, Long ottId);
 }
