+++ Back_End/src/main/java/com/ssafy/youniverse/repository/OttMemberRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.OttMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OttMemberRepository extends JpaRepository<OttMember, Integer> {
+public interface OttMemberRepository extends JpaRepository<OttMember, Long> {
 }
