+++ Back_End/src/main/java/com/ssafy/youniverse/repository/OttMovieRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.OttMovie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OttMovieRepository extends JpaRepository<OttMovie, Integer> {
+public interface OttMovieRepository extends JpaRepository<OttMovie, Long> {
 }
